Started working on module structure
