+++ src/blocks/popup/__input/popup__input.css
@@ -8,3 +8,7 @@
   font-size: 14px;
   width: 100%;
 }
+
+.popup__input:invalid {
+  border-bottom: 1px solid rgba(255, 0, 0, 1);
+}
